Effectively removes the autoGeneratedKeys feature from the driver, as if NO_GENERATED_KEYS was always set.
This prevents the driver from automatically appending a RETURNING * to an SQL statement, which the driver does even if the statement is a SELECT or already has a RETURNING clause present. This feature that was introduced with 8.4 breaks the current implementation of Adobe ColdFusion and Lucee, both of which pass on user-provided SQL without checking if NO_GENERATED_KEYS should be set, which in turn leads to broken SQL being sent to the server.

This driver should not be used for code that actually relies on autoGeneratedKeys working when the flag RETURN_GENERATED_KEYS is passed (or more specifically, flag is not NO_GENERATED_KEYS).