Effectively removes the autoGeneratedKeys feature from the driver, as if NO_GENERATED_KEYS was always set.
This feature, that was introduced with 8.4, breaks the implementation of Adobe ColdFusion 8, as the latter always requests RETURN_GENERATED_KEYS for any query whatsoever. The query parsing in the JDBC driver for such queries leads to no tuples being returned in case of an INSERT/UPDATE query containing an explicit RETURNING clause.

This driver should not be used for code that actually relies on autoGeneratedKeys working when the flag RETURN_GENERATED_KEYS is passed (or more specifically, flag is not NO_GENERATED_KEYS).